+++ CHANGELOG.md
@@ -30,6 +30,7 @@ For more information about each release including git tags and artifacts, see [R
 
 ### Fixed
 
+- Claimed-but-unactioned jobs never returning to the queue once an org had more than 128 active review queues ([#1193](https://github.com/roostorg/coop/pull/1193) by [@ThisIsMissEm](https://github.com/ThisIsMissEm))
 - Rule history dropping other rules' versions when filtered by start date ([#1056](https://github.com/roostorg/coop/pull/1056) by [@juanmrad](https://github.com/juanmrad))
 - `RetryFailedNcmecDecisionsJob` ignoring `NCMEC_ENV` and retrying test decisions ([#928](https://github.com/roostorg/coop/pull/928) by [@taobojlen](https://github.com/taobojlen))
 - Queue creation failing with "name already exists" on the default reviewer selection ([#1069](https://github.com/roostorg/coop/pull/1069) by [@jess-upscrolled](https://github.com/jess-upscrolled), closes [#1074](https://github.com/roostorg/coop/issues/1074))